refactor(sdl): extract scancode code to `scancode.cpp`
