+++ include/game/state.h
@@ -1,7 +1,7 @@
 #ifndef STATE_H
 #define STATE_H
 
-#include "platform/implementation.h"
+#include "platform/scancode.h"
 #include <cstdio>
 
 constexpr int MAX_PLAYERS = 50;
