+++ include/menu/dialog.h
@@ -1,7 +1,7 @@
 #ifndef MENU_DIALOG_H
 #define MENU_DIALOG_H
 
-#include "platform/implementation.h"
+#include "platform/scancode.h"
 
 DikScancode menu_dialog(const char* text1, const char* text2 = nullptr, const char* text3 = nullptr,
                         const char* text4 = nullptr, const char* text5 = nullptr,
