+++ include/platform/implementation.h
@@ -7,9 +7,6 @@
 
 class pic8;
 
-// DIK_ Windows scancode
-typedef int DikScancode;
-
 class palette {
     void* data;
 
@@ -37,26 +34,9 @@ bool was_right_mouse_just_clicked();
 void show_cursor();
 void hide_cursor();
 
-// Returns true if the key is currently held down.
-// Used for continious input detection like game controls
-bool is_key_down(DikScancode code);
-
-// Returns true if the key was pressed this frame (edge trigger, not held).
-// Used for single-press input detection like menu navigation
-bool was_key_just_pressed(DikScancode code);
-
-// Returns the scancode of any key pressed this frame, or DIK_UNKOWN if none.
-// Used for detecting any key press for "press any key" prompts
-DikScancode get_any_key_just_pressed();
-
-// Returns true if the key is held at OS-level key repeat intervals.
-// Used for OS-level key repeat functionality (text input)
-bool was_key_down(DikScancode code);
-
 int get_mouse_wheel_delta();
 
 std::string get_clipboard_text();
-bool is_shortcut_modifier_down();
 
 bool is_fullscreen();
 long long get_milliseconds();
