+++ src/menu/controls.cpp
@@ -4,6 +4,7 @@
 #include "main.h"
 #include "menu/nav.h"
 #include "platform/implementation.h"
+#include "platform/scancode.h"
 #include "platform/text_input.h"
 #include <cstring>
 #include <directinput/scancodes.h>
