+++ src/menu/intro.cpp
@@ -15,8 +15,8 @@
 #include "pic/pic8.h"
 #include "pic/surface.h"
 #include "platform/implementation.h"
+#include "platform/scancode.h"
 #include "renderer/render.h"
-#include <directinput/scancodes.h>
 
 static void show_intro_screen() {
     pic8* intro_screen = lock_backbuffer_pic(false);
