+++ src/menu/nav.cpp
@@ -4,13 +4,13 @@
 #include "pic/abc8.h"
 #include "pic/surface.h"
 #include "platform/implementation.h"
+#include "platform/scancode.h"
 #include "platform/text_input.h"
 #include "platform/utils.h"
 #include "util/file_iter.h"
 #include "util/util.h"
 #include <algorithm>
 #include <cstring>
-#include <directinput/scancodes.h>
 #include <numeric>
 
 menu_nav::menu_nav(std::string title)
