+++ src/platform/sdl/scancode.cpp
@@ -1,0 +1,43 @@
+#include "platform/scancode.h"
+#include "main.h"
+#include "platform/sdl/keyboard.h"
+#include <sdl/scancodes_windows.h>
+
+bool is_key_down(DikScancode code) {
+    if (code < 0 || code >= MaxKeycode) {
+        internal_error("code out of range in is_key_down()!");
+    }
+
+    SDL_Scancode sdl_code = windows_scancode_table[code];
+
+    return keyboard::is_down(sdl_code);
+}
+
+bool was_key_just_pressed(DikScancode code) {
+    SDL_Scancode sdl_code = windows_scancode_table[code];
+    return keyboard::was_just_pressed(sdl_code);
+}
+
+DikScancode get_any_key_just_pressed() {
+    for (int i = 0; i < MaxKeycode; i++) {
+        if (was_key_just_pressed(i)) {
+            return i;
+        }
+    }
+
+    return DIK_UNKNOWN;
+}
+
+bool was_key_down(DikScancode code) {
+    SDL_Scancode sdl_code = windows_scancode_table[code];
+    return keyboard::was_down(sdl_code);
+}
+
+bool is_shortcut_modifier_down() {
+    SDL_Keymod mod = SDL_GetModState();
+#ifdef __APPLE__
+    return (mod & KMOD_GUI) != 0;
+#else
+    return (mod & KMOD_CTRL) != 0;
+#endif
+}
